REF: Move order class to distinct module.
